refactor: 사용하지 않는 로그인 페이지 삭제
