--- src/app/login/actions.tsx
@@ -1,65 +1,0 @@
-'use server'
-
-import { revalidatePath } from 'next/cache'
-import { redirect } from 'next/navigation'
-
-import { createClient } from '@/utils/supabase/server'
-
-// 이메일 로그인
-export async function login(formData: FormData) {
-  const supabase = await createClient()
-  const data = {
-    email: formData.get('email') as string,
-    password: formData.get('password') as string,
-  }
-  const { error } = await supabase.auth.signInWithPassword(data)
-  if (error) {
-    redirect('/error')
-  }
-  revalidatePath('/', 'layout')
-  redirect('/')
-}
-
-// 카카오 로그인
-export async function signInWithKakao(formData: FormData) {
-  const supabase = await createClient()
-  const { data, error } = await supabase.auth.signInWithOAuth({
-    provider: 'kakao',
-    options: {
-      redirectTo: process.env.NEXT_PUBLIC_BASE_URL, // 리디렉션될 url
-    },
-  })
-  if (error) {
-    redirect('/error')
-  }
-  revalidatePath('/', 'layout')
-  redirect('/')
-}
-
-// 이메일 회원가입
-export async function signup(formData: FormData) {
-  const supabase = await createClient()
-  const data = {
-    email: formData.get('email') as string,
-    password: formData.get('password') as string,
-  }
-  const { error } = await supabase.auth.signUp(data)
-  if (error) {
-    redirect('/error')
-  }
-  revalidatePath('/', 'layout')
-  redirect('/')
-}
-
-// 로그아웃
-export async function signout() {
-  const supabase = await createClient()
-  const { error } = await supabase.auth.signOut()
-
-  if (error) {
-    redirect('/error')
-  }
-
-  revalidatePath('/', 'layout')
-  redirect('/')
-}
